backend/requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
coverage==5.2.1
ddtrace==0.41.2
Django==3.0.8
django-auto-prefetching==0.1.8
django-cacheback==2.1.0
django-cors-headers==3.4.0
django-coverage-plugin==1.8.0
django-debug-toolbar==2.2
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.10
importlib-metadata==1.7.0
intervaltree==3.1.0
kombu==4.6.11
Pillow==7.2.0
protobuf==3.13.0
PyJWT==1.7.1
python-memcached==1.59
pytz==2020.1
sentry-sdk==0.17.2
six==1.15.0
sortedcontainers==2.2.2
sqlparse==0.3.1
tenacity==6.2.0
urllib3==1.25.10
vine==1.3.0
zipp==3.1.0
